Relocate useArchive to metabase/archive too
