+++ frontend/src/metabase/archive/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./use-delete-item";
 export * from "./use-restore";
+export * from "./use-set-archive";
