+++ frontend/src/metabase/common/hooks/index.ts
@@ -15,7 +15,6 @@ export * from "./use-url-with-utm";
 export { useStoreUrl } from "./use-store-url/use-store-url";
 export * from "./use-capture-event";
 export * from "./use-progressive-loader";
-export * from "./use-set-archive";
 export * from "./use-set-collection";
 export * from "./use-set-pinned";
 export * from "./use-snapshot-selector";
