+++ frontend/src/metabase/embedding/components/ArchiveButton.tsx
@@ -2,8 +2,8 @@ import { useDisclosure } from "@mantine/hooks";
 import { c, t } from "ttag";
 
 import { useGetCollectionQuery } from "metabase/api";
+import { type ArchivableItem, useSetArchive } from "metabase/archive/hooks";
 import { canArchiveItem } from "metabase/collections/utils";
-import { type ArchivableItem, useSetArchive } from "metabase/common/hooks";
 import { ActionIcon, Icon, Menu } from "metabase/ui";
 import type { CollectionItem } from "metabase-types/api";
