+++ frontend/src/metabase/timelines/collections/containers/EditEventModal/EditEventModal.tsx
@@ -6,7 +6,7 @@ import {
   useGetTimelineQuery,
   useUpdateTimelineEventMutation,
 } from "metabase/api";
-import { useSetArchive } from "metabase/common/hooks";
+import { useSetArchive } from "metabase/archive/hooks";
 import type { ModalComponentProps } from "metabase/hoc/ModalRoute";
 import { useDispatch } from "metabase/redux";
 import EditEventModal from "metabase/timelines/common/components/EditEventModal";
