+++ frontend/src/metabase/timelines/collections/containers/EditTimelineModal/EditTimelineModal.tsx
@@ -5,7 +5,7 @@ import {
   useGetTimelineQuery,
   useUpdateTimelineMutation,
 } from "metabase/api";
-import { useSetArchive } from "metabase/common/hooks";
+import { useSetArchive } from "metabase/archive/hooks";
 import type { ModalComponentProps } from "metabase/hoc/ModalRoute";
 import { useDispatch } from "metabase/redux";
 import EditTimelineModal from "metabase/timelines/common/components/EditTimelineModal";
