+++ frontend/src/metabase/timelines/questions/containers/TimelinePanel/TimelinePanel.tsx
@@ -1,8 +1,8 @@
 import type { ComponentProps } from "react";
 
 import { useGetCollectionQuery } from "metabase/api";
+import { useSetArchive } from "metabase/archive/hooks";
 import { LoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper";
-import { useSetArchive } from "metabase/common/hooks";
 import { ROOT_COLLECTION } from "metabase/entities/collections";
 import type { CollectionId, TimelineEvent } from "metabase-types/api";
